+++ lib/upgradedb/postgres.2017101202.php
@@ -84,7 +84,8 @@ $this->Execute("
 	ALTER TABLE passwd ALTER COLUMN domainid DROP DEFAULT;
 	ALTER TABLE aliases ALTER COLUMN domainid DROP DEFAULT;
 	ALTER TABLE aliasassignments ALTER COLUMN aliasid DROP DEFAULT;
-	ALTER TABLE aliasassignments ALTER COLUMN accountid DROP DEFAULT;
+	ALTER TABLE aliasassignments ALTER COLUMN accountid DROP NOT NULL;
+	ALTER TABLE aliasassignments ALTER COLUMN accountid SET DEFAULT NULL;
 	ALTER TABLE eventassignments ALTER COLUMN eventid DROP DEFAULT;
 	ALTER TABLE daemoninstances ALTER COLUMN hostid DROP DEFAULT;
 	ALTER TABLE daemonconfig ALTER COLUMN instanceid DROP DEFAULT;
@@ -321,13 +322,13 @@ if (empty($ids)) {
 	$this->Execute("DELETE FROM aliases WHERE domainid NOT IN (" . $sql_ids . ")");
 }
 
-$this->Execute("DELETE FROM aliasassignments WHERE accountid = 0");
+$this->Execute("DELETE FROM aliasassignments WHERE accountid = 0 AND mail_forward=''");
 $ids = $this->GetCol("SELECT id FROM passwd");
 if (empty($ids)) {
-	$this->Execute("DELETE FROM aliasassignments");
+	$this->Execute("DELETE FROM aliasassignments WHERE mail_forward=''");
 } else {
 	$sql_ids = implode(',', $ids);
-	$this->Execute("DELETE FROM aliasassignments WHERE accountid NOT IN (" . $sql_ids . ")");
+	$this->Execute("DELETE FROM aliasassignments WHERE mail_forward='' AND accountid NOT IN (" . $sql_ids . ")");
 }
 
 $this->Execute("DELETE FROM aliasassignments WHERE aliasid = 0");
